Rename bench: keys.rs -> post_keys.rs
