+++ nexus-webapi/benches/streams_benches/mod.rs
@@ -1,7 +1,7 @@
 pub mod author;
 pub mod bookmarks;
-pub mod keys;
 pub mod kind;
+pub mod post_keys;
 pub mod reach;
 pub mod sorting;
 pub mod tag;
